+++ EShop.Domain/SalePoint.cs
@@ -1,4 +1,4 @@
-﻿namespace EShop.Presentation
+﻿namespace EShop.Domain
 {
     public class SalePoint
     {
