+++ EShop.Presentation/EShop.Presentation.http
@@ -1,6 +1,4 @@
 @EShop.Presentation_HostAddress = http://localhost:5027
 
-GET {{EShop.Presentation_HostAddress}}/weatherforecast/
-Accept: application/json
 
 ###
