+++ EShop.Presentation/SalePointDto.cs
@@ -1,0 +1,8 @@
+﻿namespace EShop.Presentation
+{
+    public record SalePointDto
+    {
+        public string Name { get; set; }
+        public int SaleArea { get; set; }
+    }
+}
