+++ EShop.sln
@@ -1,9 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35707.178 d17.12
+VisualStudioVersion = 17.12.35707.178
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EShop.Presentation", "EShop.Presentation\EShop.Presentation.csproj", "{911A9F4E-0A89-4AC7-834A-D9FEF43D4C56}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EShop.Domain", "EShop.Domain\EShop.Domain.csproj", "{624E1CEF-F87A-45F2-80EB-85B05695C53F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EShop.Application", "EShop.Application\EShop.Application.csproj", "{4E179B39-3CBD-4EA0-95DD-C2ABBD70467B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +19,14 @@ Global
 		{911A9F4E-0A89-4AC7-834A-D9FEF43D4C56}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{911A9F4E-0A89-4AC7-834A-D9FEF43D4C56}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{911A9F4E-0A89-4AC7-834A-D9FEF43D4C56}.Release|Any CPU.Build.0 = Release|Any CPU
+		{624E1CEF-F87A-45F2-80EB-85B05695C53F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{624E1CEF-F87A-45F2-80EB-85B05695C53F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{624E1CEF-F87A-45F2-80EB-85B05695C53F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{624E1CEF-F87A-45F2-80EB-85B05695C53F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4E179B39-3CBD-4EA0-95DD-C2ABBD70467B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4E179B39-3CBD-4EA0-95DD-C2ABBD70467B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4E179B39-3CBD-4EA0-95DD-C2ABBD70467B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4E179B39-3CBD-4EA0-95DD-C2ABBD70467B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
